Update README.md
The Temperature Converter Webpage is a simple and user-friendly tool for converting temperatures between Celsius, Fahrenheit, and Kelvin. This repository provides all the necessary files, including HTML, CSS, and JavaScript, to run the converter on any web browser. By entering a temperature value and selecting the input and output units, users can quickly obtain the converted temperature with just a click of a button. The webpage features an attractive design with a responsive layout and a visually appealing background image. It is easy to navigate, and the clear instructions guide users through the conversion process. Whether you want to convert Celsius to Fahrenheit, Fahrenheit to Kelvin, or any other temperature conversion, this webpage provides an efficient and convenient solution. The project is open to contributions, and the code is released under the MIT License, allowing users to customize and modify it according to their needs. Enjoy using the Temperature Converter Webpage to simplify your temperature conversions!